vm-runner: use multi-queue networking
Enable multiqueue support for virtio-net-pci devices as we seen them choking on traffic and dropping
packets. Set queues=4 and vectors=10 as a reasonable default. `vectors` should be to 2*queues + 2 as
per https://www.linux-kvm.org/page/Multiqueue. We also enable multiqueue support for all VM sizes, even
to a small ones. As of time of writing, it seems to not worth a trouble to dynamically adjust number
of queues based on VM size.