agent: Add support for simplified LFC scaling
We have one ongoing issue where the LFC goal size algorithm is hitting
degenerate edge-case behavior on a large workload -- but because of that
bug, it's hard to tell whether the impacted workload *can* be feasibly
autoscaled with the information we have available.

For large workloads, it's reasonable enough to just look at the 1-hour
working set window, and testing with that will indicate whether we need
to look at a longer time horizon to effectively scale LFC on such a
workload -- or if the information we already have should be sufficient.

See also neondatabase/cloud#22214 for more information.