agent: Decrease LFC metrics fetch frequency from 1/15hz -> 1/60hz
I.e., instead of fetching LFC metrics every 15 seconds, we will now
fetch them every 60s.

However, going up to a minute without LFC metrics wouldn't be good,
because we require all sources of metrics to be available to downscale.
So this commit adds new functionality to the metrics fetching config to
guarantee that the *first* metrics fetch is done within a smaller time
interval, even if the rest are evenly distributed over the full range of
the refresh period.