neonvm: mv root disk from init container, not cp (#1189)
This commit is relatively straightforward: In the neonvm-runner init
container, switch from copying the root disk into `/vm/images` to
*moving* it there.

Copying necessarily causes a lot of disk writes, but moving may not. If
the init container's `disk.qcow2` is on the same underlying filesystem
as the `/vm/images` volume, then the kernel MAY do the move very
efficiently.

To be clear: I'm not totally sure whether this will help — it may be
that moving files out of an overlayfs still requires copying. But if it
helps, it could dramatically reduce our disk IO usage when many computes
are started at once. And even if it doesn't help, it's probably a better
default to move instead of copying.