Add cluster-trace-collector tool
This new tool exists so that we can capture a scheduling/scaling "trace"
from production (basically: a record of exactly all the operations that
happened). This way, we have some real data availalbe to use in
simulations for changes to our scheduling policies, AND we can use that
data to extrapolate out to how changes in load will be handled.

The architecture is pretty simple:

* Reuse pkg/util/watch to get a stream of events on Pod and Node objects
* Extract the useful bits out of each event
* If the useful bits changed, send it off to S3/Azure Blob with an event
  sink from pkg/reporting