agent: Add per-VM metric for desired CU(s) (#1108)
This commit adds a new per-VM metric: autoscaling_vm_desired_cu.

It's based on the same "desired CU" information exposed by the scaling
event reporting, but updated continuously instead of being rate limited
to avoid spamming our reporting.

The metric has the same base labels as the other per-VM metrics, with
the addition of the "component" label, which is one of:

* total - the goal CU, after taking the maximum of the individual parts
  and rounding up to the next unit.
* cpu - goal CU size in order to fit the current CPU usage
* mem - goal CU size in order to fit the current memory usage
  (including some information derived from LFC, to make sure there's
  room for cache too)
* lfc - goal CU size in order to fit the estimated working set size

All of these values are also multiplied by the same Compute Unit
multplier as with the normal scaling event reporting, so that Neon's
fractional compute units are exposed as such in the metrics, even as we
use integer compute units in the autoscaler-agent.

Also note that all values except "total" are NOT rounded, and instead
show the fractional amounts to allow better comparison.