Allow neonvm-daemon to handle resize-swap and set-disk-quota
Introduces a new internal-facing interface to neonvm-daemon, which can
be used to resize the swap disk and to set the disk quota on the
pgdata disk.

This will replace the sudo scripts that compute_ctl currently
uses. They are kept around for now for backwards-compatibility, but
when the compute image builds have switched over to using the new
interface, they can be removed.

One noteworthy thing is that the path controlled by the quota is now
fixed, to /var/db/postgres/compute. It's a CLI argument to
neonvm-daemon, but the inittab always passes /var/db/postgres/compute.
This could be parameterized further if needed, but I don't see the
need at the moment.

Adds a README explaining neonvm-daemon's role, and the two interfaces
it now has.