plugin/metrics: Update metrics in-place when possible (#1277)
Fixes #1276.

Currently, the way we update these node metrics is by removing all the
old ones, then adding back the current values. We do it that way so that
the old values can be cleaned up when there's label changes.

However: if metrics are scraped in between removing the old and adding
the new, we can end up with single-datapoint gaps for one node at a
time.

So to fix this, we should avoid removing the old metrics if and only if
the labels are unchanged -- which we can check just by storing the
previous labels we used.