reporting: Serialize batches as they're constructed (#1227)
Part of #1220.

For scaling events, the size of a gzip-compressed batch is *much* less
than the uncompressed version, so it's actually worth it to compress the
JSON-encoded contents of the batch as we go, instead of buffering the
events as a []ScalingEvent in memory.

**Implementation**

This change primarily adds the reporting.BatchBuilder[E] type (and some
implementations), which abstracts over the process of gradually building
the serialized bytes from a batch of events.

Two implementations of BatchBuilder are provided in pkg/reporting - one
for large JSON arrays, and one for the JSON lines format.

To allow continuous gzip compression, both of these JSON batch builders
are constructed with a new 'IOBuffer' type, which is just an io.Writer
with a method to collect the bytes at the end. There's implementations
of IOBuffer for normal bytes and for a gzip-compressed buffer.

All that together means that batches of billing and scaling events
should now be GZIP-compressed (where possible) as events are added to
the batches, dramatically reducing the memory footprint.