ci/build-images: Standardize image sets (#1312)
Basically, make there's the same set of images every time and they're in
the same order. And if there's an image intentionally missing, comment
why.

Also, align spacing in the places where that's feasible, because it
makes it easier to see the set of images.

---

Noticed some painful merge conflicts trying to update #1247, figured
this bit of cleanup would help (it does!).