go.mod: Update k8s deps to 1.31.7
Also update cert-manager, so that its dependencies are also kept in
sync.

The only straggler is sigs.k8s.io/apiserver-network-proxy/konnectivity-client,
which is at v0.30.3 even though it's versioned to match the kubernetes
version. HOWEVER that's the version that k/k itself uses (and also what
various k8s-sigs projects use). So presumably it's fine.