plugin/state: Include "committed" resources in metrics (#1337)
Follow-up to #1289, increasing observability.
Part of neondatabase/cloud#25458.

This commit adds a new field to the plugin's state.NodeResources to
store the amount of resources "committed" -- i.e. the amount before
applying any overcommit factors.

Then, surfacing it in the metrics is as simple as adding the field to
the list from (NodeResources[T]).Fields().