neonvm: GZIP-compress base64 runner args
For VirtualMachine objects with very large spec fields, we see that the
base64 encoded version, which is passed as an argument to neonvm-runner,
massively contributes to the total size of the resulting pod object.

Because of the high rate of updates to pods, this can have additional
large effects on the total amount of data stored in etcd.

Long-term I would like us to move away from packing the entire spec and
status into the neonvm-runner args, but that's a much larger project.

See INC-492 for more:
https://neondb.slack.com/archives/C08N5D2CF7A/p1744380652602419