CI: GHCR as the main registry (#1283)
We're shifting the main workload from Docker Hub to GHCR.

This PR makes CI in the repo use GHCR as the main container registry.
For releases (`inputs.upload-to-ecr-acr` is set to `true`) we still
upload images to Docker Hub for compatibility.

This PR also replaces `_push-to-acr.yaml` workflow with
`_push-to-container-registry.yml` (which we already use in `neon` repo
to spread docker images across registries)