neonvm-controller: Add flags for k8s client + leader election config (#1383)
We'd like to try different leader election configuration in order to see
if that helps with some issues we're seeing in production. Alternately,
we may want to adjust our client rate limits.

In order to facilitate quick experimentation, we should use command line
flags, so that we don't need cut a fresh release each time we want to
try something new.

ref https://neondb.slack.com/archives/C08SNRE78UQ/p1747302099668289