neonvm: Remove neonvm-runner version [1/2] (#1380)
We no longer use it anywhere, so we should probably remove it to get rid
of unnecessary code.

This PR is the first of two changes, removing only the parts of
neonvm-controller that check the version label while continuing to set
the label for new neonvm-runner pods. It must be two separate changes to
ensure rollback safety.

See #1381 for more.