go: add syntax for error handling to Neon Go
After a recent announcement [1] by Go team, saying they won't proceed
with adding error handling syntax to Go, it is evident we have to
implement a solution downstream.

We already build Linux kernel with custom patches (see #1386, #1392), it
is logical that as a next step we should have our own fork of Go, called
Neon Go.

The first feature is a syntax to handle errors. It supports the
following syntax:

  ... := ... handle err {
    ...
  }

Which is equivalent to:

  ..., err := ...
  if err != nil {
    ...
  }

Advantages:
1. Easy to read.
2. Easy to write (no need to look for the ! key).

Timeline to implement:
5 June - merge to main
6 June - deploy to staging
7 June - deploy to production

[1]: https://go.dev/blog/error-syntax

NOTE: this is, of course, a joke. I don't have any strong opinion about
error handling in Go.