neonvm-kernel: Patch null ptr deref in x86 pagetable setup (#1392)
ref https://neondb.slack.com/archives/C0807C9SSJ2/p1748363834798349
ref https://neondb.slack.com/archives/C03TN5G758R/p1748944236243709
ref #1391

Broader discussion in the slack threads, but at a high level:

1. We started seeing null pointer dereferences after failed allocations
   in phys_pud_init() and friends
2. Those functions don't handle allocation failure (because normally
   they're called during boot, where the allocation cannot fail)
3. The allocation uses the GFP_ATOMIC flag, which *can* fail
4. This has been the case for a while, so whatever reason we're starting
   to see allocation failures, this code probably does need to be able to
   handle them.

This change consists of several patches, matching this thread on LKML:
https://lore.kernel.org/all/7d0d307d-71eb-4913-8023-bccc7a8a4a3d@neon.tech/

The main idea of the change is to propagate allocation failures by
changing the phys_*_init() functions to return -ENOMEM (which in
turn requires changing their return type).

This patch was tested by injecting failure in phys_pud_init() for a
single call to alloc_low_page() after boot, confirming that:

* an allocation failure will no longer cause a null pointer dereference
* virtio-mem hotplug succeeds after the initial failure, if made to
  retry

Check the slack threads above for more detail on that testing.