neonvm-kernel: Add debug logs for kcompactd/offlining race (#1400)
We are currently seeing a small base rate of a kernel bug where
kcompactd is unable to handle a page fault during compaction.

We suspect this *may* be due to a race condition between memory
offlining and kcompactd, but it's hard to establish that concretely
without more logs.

So, let's add some debug logs indicating when kcompactd runs or
offlining happens.

ref https://neondb.slack.com/archives/C0807C9SSJ2/p1749652291036669