neonvm: Remove neonvm-runner version [2/2] (#1381)
We no longer use it anywhere, so we should probably remove it to get rid
of unnecessary code.

This commit is the second of two changes, removing the remaining places
where we set the neonvm-runner version label. It must be a separate
change so that neonvm-controller can first be updated to ignore whether
the label is present, for rollback safety.