Support multi-arch image SHAs (#1407)
When building both x86 and ARM images on the same machine, docker can
get confused if you end up building the same image twice, giving an
error like:

  cannot overwrite digest sha256:e5d0aea7f7d2954678a9a6269ca2d06e06591881161961ea59e974dff3f12377

This happens because pulling a multi-platform image using the SHA of the
*manifest* will resolve to different underlying images.

To fix this, we need to use the platform-specific digests, which we can
get from the 'docker manifest inspect' output of the multi-platform
digest.

---

To test on apple silicon with rosetta enabled, you can run:

  make HOST_ARCH=arm64 kernel docker-build
  make HOST_ARCH=amd64 kernel docker-build

(i.e. with both architectures)