agent,plugin: Cancel orca ctx before wait
In LKB-2414, we found that the autoscaler-agent can deadlock after a
failure during startup. This is because the deferred call to
srv.GetOrchestrator().Wait() is *before* the canceling of the context,
which means that it just blocks forever.

To fix the immediate issue, move the cancel() into the same deferred
function.