neonvm-controller: fix a bug of migration state machine (#1427)
The state machine management in
neonvm-controller might have an issue. **In the `PreMigrating` state,
the phase is always transitioned directly into `Migrating`, regardless
of the actual result of the QMPrequest.**

Most of the time this isn’t a problem, but in some corner cases where
the target QEMU startup is slower than expected and the QMP port becomes
available later, this can cause a request failure.

When that happens, neonvm-controller doesn’t perform any retries—it
simply moves into the `Migrating` phase, which causes the migration flow
to get stuck. After applying the change in this PR, that issue is
resolved.

Co-authored-by: Oleg Vasilev <oleg@neon.tech>